refactor(history): split HistoryResultList into explicit topic/session variants
The result list used a `mode: 'assistant'|'session'` discriminator plus 8
parallel topic/session prop pairs (16 props), with only one half live per
mode and the render branched into two fully separate row subtrees. The sole
consumer passed dead `EMPTY_*_BY_ID`/empty-array fillers for the unused half.

Replace it with two explicit variant components over a shared generic table:
- HistoryTable.shared.tsx — the generic HistoryVirtualTable, header, cells,
  context menu, action/pin/delete buttons, and time formatter.
- HistoryTopicResultList.tsx — topic-only props + row (estimateSize 44).
- HistorySessionResultList.tsx — session-only props + row (estimateSize 52),
  dropping the dead unlinkedAssistantLabel.

HistoryRecordsPage renders each variant directly, dropping the discriminator
and the empty-map fillers. DOM-identical → the page tests pass unchanged.

Also update the AgentComposer test mock for the renderSendAccessory →
sendAccessory slot rename.
